requirements.txt
scrapy
pillow